Refactored to Header, Left Panel and Right Panel. Still much work to be done
