+++ senatus/src/extensions/documentation/documentation/1.0.0/full_documentation.json
@@ -14,7 +14,7 @@
       "name": "Apache 2.0",
       "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
     },
-    "x-generation-date": "2026-04-08T17:39:41.019Z"
+    "x-generation-date": "2026-04-11T10:13:47.181Z"
   },
   "x-strapi-config": {
     "plugins": [
