+++ web/src/pages/CompetePage/components/ErrorComponent.tsx
@@ -1,0 +1,7 @@
+export default function ErrorComponent(){
+  return(
+    <div className="text-center py-12">
+      <p className="text-destructive">Failed to load problem. Please try again later.</p>
+    </div>
+  )
+}
