+++ web/src/pages/CompetePage/components/Loading.tsx
@@ -1,0 +1,7 @@
+export default function Loading(){
+  return (
+    <div className="flex items-center justify-center h-[calc(100vh-8rem)]">
+      <div className="animate-spin rounded-full h-8 w-8 border-2 border-primary border-t-transparent"></div>
+    </div>
+  )
+}
