+++ web/src/pages/CompetePage/components/ProblemMode/ProblemMode.tsx
@@ -1,0 +1,12 @@
+import TrainingMode from "./TrainingMode";
+import ViewMode from "./ViewMode";
+
+export default function ProblemMode({ isTrainingMode, isViewMode }: { isTrainingMode: boolean, isViewMode: boolean }) {
+  return (
+    <>
+      <ViewMode isViewMode={isViewMode} />
+      <TrainingMode isTrainingMode={isTrainingMode} />
+    </>
+
+  )
+}
