+++ web/src/pages/CompetePage/types/EventItem.ts
@@ -1,0 +1,8 @@
+import { EventProblem } from "./EventProblem";
+
+export interface EventItem {
+  documentId: string;
+  title: string;
+  end: string;
+  problems?: EventProblem[];
+}
