+++ web/src/pages/CompetePage/types/EventProblem.ts
@@ -1,0 +1,6 @@
+export interface EventProblem {
+  documentId: string;
+  title: string;
+  points?: number;
+  difficulty?: string;
+}
