+++ web/src/pages/CompetePage/types/Execution.ts
@@ -1,0 +1,16 @@
+export interface Execution {
+  id?: string;
+  documentId: string;
+  stdout: string;
+  stderr: string;
+  executionTime: number;
+  processed: boolean;
+  passed?: boolean;
+  testCase?: {
+    documentId: string;
+    input: string;
+    output: string;
+    hidden: boolean;
+    locked?: boolean;
+  };
+}
