+++ web/src/pages/CompetePage/types/TestCase.ts
@@ -1,0 +1,9 @@
+export interface TestCase {
+  documentId: string;
+  input: string;
+  output: string;
+  hidden: boolean;
+  locked: boolean;
+  weight: number;
+  explanation?: string;
+}
